% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreInteractions.R
\name{scoreInteractions}
\alias{scoreInteractions}
\alias{scoreInteractions,InteractionArray,missing-method}
\title{Calculate loop enrichment over background.}
\usage{
scoreInteractions(
  x,
  hicFile,
  loopCalls,
  fgSize = 0,
  bg = "bowtie",
  bgGap = 0,
  bgSize = 5,
  pseudo,
  pruneUnder = 0,
  truncate = F,
  resolution,
  norm = "KR"
)

\S4method{scoreInteractions}{InteractionArray,missing}(
  x,
  loopCalls,
  fgSize,
  bg,
  bgGap,
  bgSize,
  pseudo,
  pruneUnder,
  truncate
)
}
\description{
Calculate loop enrichment over background.
}
