% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopcity-package.R
\docType{package}
\name{loopcity-package}
\alias{loopcity}
\alias{loopcity-package}
\title{loopcity: Call Chromatin Communities From Hi-C Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for assigning and visualizing chromatin community interactions from Hi-C data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sarmapar.com/loopcity/}
  \item \url{https://github.com/sarmapar/loopcity}
  \item Report bugs at \url{https://github.com/sarmapar/loopcity/issues}
}

}
\author{
\strong{Maintainer}: Sarah Parker \email{smp3800@gmail.com} (\href{https://orcid.org/0000-0002-2700-3979}{ORCID})

Authors:
\itemize{
  \item Sarah Parker \email{smp3800@gmail.com} (\href{https://orcid.org/0000-0002-2700-3979}{ORCID})
}

Other contributors:
\itemize{
  \item Howard Hughes Medical Institute [funder]
  \item National Institutes of Health [funder]
  \item National Science Foundation [funder]
}

}
\keyword{internal}
