% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeAnchors.R
\name{mergeAnchors}
\alias{mergeAnchors}
\title{Merge neighboring loop anchor positions into one
 representative anchor.}
\usage{
mergeAnchors(loops, pixelOverlap = 1, dropDups = TRUE)
}
\arguments{
\item{loops}{GInteractions object of loops from Hi-C data}

\item{pixelOverlap}{numeric representing the number of
pixels away from an anchor is considered to be overlapping
For example, `pixelOverlap = 1` means only immediately neighboring pixels
will be merged and `pixelOverlap = 2` means all anchors within 2 pixels will
be merged}

\item{dropDups}{logical determining whether duplicates should be dropped
from output}
}
\value{
A GInteractions object with loop anchors merged into representative
positions.
}
\description{
The representative anchor is chosen as the mode of a set of
 neighboring anchors. If there are no modes, the middle-most anchor of the
 set of neighboring anchors is chosen. If there are multiple modes, the
 middle-most anchor of the set of modes is chosen. In the case of an even
 number in the set, the anchor at a higher genomic coordinate is chosen to
 break ties.
}
\examples{
if (requireNamespace("loopcityData", quietly = TRUE)) {
    data(GM12878_10KbLoops, package = "loopcityData")
    mergeAnchors(loops = GM12878_10KbLoops, pixelOverlap = 1)
}
}
