% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignCommunities.R
\name{assignCommunities}
\alias{assignCommunities}
\title{Assign communities to interactions}
\usage{
assignCommunities(
  loops,
  scores,
  clusterType = "leiden",
  leidenResolution = 0.1,
  pruneUnder
)
}
\arguments{
\item{loops}{GInteractions object of interactions to include in the network}

\item{scores}{a numeric vector of scores to use as edge weights,
if `loops` contains a column named 'score', this column will be used}

\item{clusterType}{character vector of length 1 containing a keyword
for the clustering algorithm to use. Must be one of "fast_greedy",
"walktrap", "leiden", "infomap", "label_prop", or "edge_betweenness"}

\item{leidenResolution}{numeric between 0 and 1, higher resolutions lead to
more smaller communities, while lower resolutions lead to fewer larger
communities. See `resolution_parameter` in `igraph::cluster_leiden`}

\item{pruneUnder}{numeric, added loops with scores less than this value will
be removed}
}
\value{
GInteractions object with community metadata
}
\description{
Assign communities to interactions
}
\examples{
data(GM12878_10KbLoopsChr21)
loops <- GM12878_10KbLoopsChr21
loops$score <- loops$value
assignCommunities(loops, pruneUnder = 0)
}
