% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignCommunities.R
\name{assignCommunities}
\alias{assignCommunities}
\title{Assign communities to interactions}
\usage{
assignCommunities(
  loops,
  scores,
  clusterType = "leiden",
  leiden_resolution = 0.01,
  pruneUnder
)
}
\arguments{
\item{loops}{GInteractions object of interactions to include in the network}

\item{scores}{a numeric vector of scores to use as edge weights,
if `loops` contains a column named 'score', this column will be used}

\item{clusterType}{character vector of length 1 containing a keyword
for the clustering algorithm to use. Must be one of "fast_greedy",
"walktrap", "leiden", "infomap", "label_prop", or "edge_betweenness"}

\item{pruneUnder}{numeric, all loops with scores less than this value will be
removed}
}
\value{
GInteractions object with community metadata
}
\description{
Assign communities to interactions
}
\examples{
# Add example
}
