% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopcity-package.R
\docType{package}
\name{loopcity-package}
\alias{loopcity}
\alias{loopcity-package}
\title{loopcity: Call Chromatin Communities From Hi-C Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for assigning and visualizing chromatin community interactions from Hi-C data.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.sarmapar.com/loopcity/}
}

}
\author{
\strong{Maintainer}: Sarah Parker \email{smp3800@gmail.com} (\href{https://orcid.org/0000-0002-2700-3979}{ORCID})

}
\keyword{internal}
