% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHicCommunities.R
\name{plotHicCommunities}
\alias{plotHicCommunities}
\title{Plot Hi-C data with loops colored by community assignment.}
\usage{
plotHicCommunities(
  pdfName,
  communities,
  hicFile,
  norm = "SCALE",
  chroms,
  starts,
  ends,
  zmax,
  colorPalette
)
}
\arguments{
\item{pdfName}{a character string giving the file path. See the `file`
parameter in `pdf()` for more information.}

\item{communities}{GInteractions object containing loops to plot and a
column called `loopCommunity` containing community assignments}

\item{hicFile}{path to .hic file to plot}

\item{chroms}{chromosome(s) of the region(s) to plot}

\item{starts}{starting genomic location(s) of the region(s) to plot}

\item{ends}{ending genomic location(s) of the region(s) to plot}

\item{zmax}{a numeric value specifying the maximum range of hic scores to
plot, where values greater than `zmax` will be set to`zmax`}

\item{colorPalette}{a function corresponding to a color palette}
}
\description{
If a vector of starts and ends is provided, one page will be generated
per region.
}
\examples{
add example
}
