% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreInteractions.R
\name{scoreInteractions}
\alias{scoreInteractions}
\title{Score loop count using enrichment of foreground over background}
\usage{
scoreInteractions(
  mats,
  fgDist = 0,
  bg = "bowtie",
  bgDist = 5,
  pseudo = 0,
  pruneUnder = 0,
  truncate = F
)
}
\arguments{
\item{mats}{matrix of HiC count interactions.
See also `mariner::pixelsToMatrices`}

\item{fgDist}{manhattan distance surrounding center pixel to
include in foreground selection
See also parameter `n` in `mariner::selectRadius`}

\item{bg}{character keyword for background selection
One of `c("bowtie", "donut")`}

\item{pseudo}{number of pseudo counts to add to raw Hi-C counts}

\item{pruneUnder}{numeric, all scores under this value will be set to 0}

\item{truncate}{logical indicating whether to set any values above
the 99th percentile to the 99th percentile value}

\item{interactions}{GInteraction object of interactions from Hi-C data}
}
\description{
Score loop count using enrichment of foreground over background
}
